UserRepository and Test Class Creation.

+++ src/main/java/com/market/entities/User.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
-import java.util.UUID;
 
 @Entity(name = "users")
 @Table(name = "users")
@@ -12,7 +11,7 @@ import java.util.UUID;
 public class User implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String phone;
@@ -23,7 +22,7 @@ public class User implements Serializable {
 
     }
 
-    public User(int id, String name, String email, String phone, String password) {
+    public User(Long id,String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -32,11 +31,12 @@ public class User implements Serializable {
 
     }
 
-    public int getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
